requirements.txt
requests==2.8.1
PyYaml==3.11
zadarest>0.1
